+++ pytket/binders/passes.cpp
@@ -970,14 +970,19 @@ PYBIND11_MODULE(passes, m) {
       "\n:param allow_zzphase: If set to True, allows the algorithm to "
       "implement 2-qubit rotations using ZZPhase gates when deemed "
       "optimal. Defaults to False."
-      "\n:param timeout: Sets maximum out of time spent finding solution."
+      "\n:param thread_timeout: Sets maximum out of time spent finding a "
+      "single solution in one thread."
       "\n:param only_reduce: Only returns modified circuit if it has "
       "fewer two-qubit gates."
+      "\n:param trials: Sets maximum number of found solutions. The "
+      "smallest circuit is returned, prioritising the number of 2qb-gates, "
+      "then the number of gates, then the depth."
       "\n:return: a pass to perform the simplification",
       py::arg("discount_rate") = 0.7, py::arg("depth_weight") = 0.3,
       py::arg("max_lookahead") = 500, py::arg("max_tqe_candidates") = 500,
       py::arg("seed") = 0, py::arg("allow_zzphase") = false,
-      py::arg("timeout") = 100, py::arg("only_reduce") = false);
+      py::arg("thread_timeout") = 100, py::arg("only_reduce") = false,
+      py::arg("trials") = 1);
   m.def(
       "PauliSquash", &PauliSquash,
       "Applies :py:meth:`PauliSimp` followed by "